requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
multidict==4.5.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dotenv==0.9.1
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
urllib3==1.24.1
websockets==6.0
Werkzeug==0.14.1
yarl==1.2.6
-e ./discord-api-build/
